Remove authorization unit tests
This removes all authorization unit tests. From now on authorization is
not handled in models. Model does not have enough information to check
it anyway, e.g. when you clone some resource a clone permission should be
enough. On model level you just know that someone is trying to create a
record so you check for create permission which is wrong.

If we ever wanted to add similar check we would write these tests
differently anyway. These tests were completely unDRY. We would write a
module (similar to Authorization module that is being removed) and cover
this module by tests. Then we would only make sure that all models
include this module. Checking of module functionality in every model
seems as huge amount of duplication.

Since permissions are not enforced on model layer, writing unit tests
should be much easier from now on. You don't have to setup different
models to test just one class (which is the key for unit tests)